requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-Limiter
PyJWT
python-dotenv
email-validator
requests
